Create new publishing-api comparison service
This needs to raise a different error from `WhitehallImporter::AbortImportError`
