+++ app/services/resync_service.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
-require 'damerau-levenshtein'
 
 class ResyncService < ApplicationService
   attr_reader :document
@@ -24,7 +23,7 @@ private
   def sync_live_edition
     live_edition.lock!
     set_political_and_government(live_edition)
-    compare_with_publishing_api(live_edition, published: true)
+    PublishingApiComparisionService.call(live_edition, published: true)
     reserve_path(live_edition.base_path)
     PreviewService.call(live_edition, republish: true)
     PublishAssetService.call(live_edition, nil)
@@ -41,7 +40,7 @@ private
   def sync_draft_edition
     current_edition.lock!
     set_political_and_government(current_edition)
-    compare_with_publishing_api(current_edition)
+    PublishingApiComparisionService.call(current_edition)
     reserve_path(current_edition.base_path)
     FailsafePreviewService.call(current_edition)
 
@@ -109,51 +108,5 @@ private
     scheduling = current_edition.status.details
     ScheduledPublishingJob.set(wait_until: scheduling.publish_time)
                           .perform_later(current_edition.id)
-  end
-
-  def compare_with_publishing_api(edition, published: false)
-    proposed_edition = PreviewService::Payload.new(edition, republish: published).payload
-    edition_in_publishing_api = GdsApi.publishing_api.get_content(edition.content_id).to_h
-
-    if published
-      version = edition_in_publishing_api["state_history"].select { |_, hash| hash["published"] }
-      published_version = version.keys.first.to_i
-
-      edition_in_publishing_api = GdsApi.publishing_api.get_content(edition.content_id, version: published_version).to_h
-    end
-
-    raise WhitehallImporter::AbortImportError, "Versions don't match" unless versions_match?(edition_in_publishing_api, proposed_edition)
-    raise WhitehallImporter::AbortImportError, "Links don't match" unless links_match?(content_id, proposed_edition)
-  end
-
-  def versions_match?(edition_in_publishing_api, proposed_edition)
-    edition_in_publishing_api["base_path"] == proposed_edition["base_path"] &&
-      edition_in_publishing_api["title"] == proposed_edition["title"] &&
-      edition_in_publishing_api["description"] == proposed_edition["description"] &&
-      edition_in_publishing_api["first_published_at"] == proposed_edition["first_published_at"] &&
-      edition_in_publishing_api["public_updated_at"] == proposed_edition["public_updated_at"] &&
-      edition_in_publishing_api["document_type"]["id"] == proposed_edition["document_type"] &&
-      edition_in_publishing_api["schema_name"] == proposed_edition["schema_name"] &&
-      details_match?(edition_in_publishing_api["details"], proposed_edition["details"])
-  end
-
-  def details_match?(pub_api_details, proposed_details)
-    body_text_similar_enough?(pub_api_details["body"], proposed_details["body"]) &&
-    images_match?()
-  end
-
-  def body_text_similar_enough?(pub_api_body, proposed_body)
-    # See https://www.rubydoc.info/gems/damerau-levenshtein/1.1.0#API_Description
-    DamerauLevenshtein.distance(pub_api_body, proposed_body, 1, 100) < 100
-  end
-
-  def links_match?(content_id, proposed_edition)
-    links_in_publishing_api = GdsApi.publishing_api.get_links(content_id).to_h
-    proposed_links = proposed_edition["links"]
-
-    links_in_publishing_api["government"].sort == proposed_links["government"].sort &&
-      links_in_publishing_api["organisations"].sort == proposed_links["organisations"].sort &&
-      links_in_publishing_api["taxons"].sort == proposed_links["taxons"].sort &&
-      links_in_publishing_api["primary_publishing_organisation"] == proposed_links["primary_publishing_organisation"]
   end
 end
